Revert "Remove zsh-histdb"
This reverts commit a46b8a5506d617fe27336358c6e348da9a1da1b7.

Switch back to using zsh-histdb and discard mcfly. Now that I have
enabled ^R and customised the search widget, zsh-histdb feels more
usable. McFly was buggy, failing on multi-line commands, and switching
to the alt screen was a usability issue for me. Seeing the history was
nice, but blanking the working terminal means the context is lost.

Signed-off-by: Donald Hunter <donald.hunter@gmail.com>